organizing code and removing logs
